// src/types/package.ts
export type Package = {
  no: number;
  namaPelanggan: string;
  namaPerangkat: string;
  jumlahAir: number;
  keterangan: string;
};

export type Package2 = {
  no: number;
  namaPelanggan: string;
  namaPerangkat: string;
  jumlahAir: number;
  totTagih: number;
  tenggat: string;
  status: string;
};